Add "Add Parallax Series" form
(In the background tab) Double click and drag the mouse around to select the area, let go of the mouse, and fill in the desired information in the form. From there, the selected area will be filled with new parallax lines.

(this feature is mostly for stuff ie the GHZ parallax lines, where there are like 100 parallax lines right after one another with like a 0.003 change in parallax factor every line)