Add (untested) fix for 16x16Tiles.gif corruption on Mono
As said in the code comments, the code itself works, it's just whether or not it'll fix the issue on Linux that's unknown. Seemingly, the issue stems from the normal GDI method not being supported properly through Wine/Mono? So, swapping it out for the standard RSDKv3_4.Gif class *should* work, although this method is slower.

I've been sitting on this code for a while, figured it's probably best to push it sooner rather than later in the case that maybe it does actually work?- And if it doesn't, then.. sorryyyy